CLean up and restructure framework
